Optimize: Pruned redundant code in R_RegisterSkin; we now use the resource locator for searching a model's local directory and will also handle trying each supported graphic file format in said location.

+++ doomsday/engine/portable/include/r_model.h
@@ -74,7 +74,7 @@ typedef struct {
     char            skinRange;
     float           offset[3];
     byte            alpha;
-    short           shinySkin; // Skinname ID (index)
+    uint            shinySkin; // Skinname ID (index)
 } submodeldef_t;
 
 #define MODELDEF_ID_MAXLEN      32
